gce: initial IP alias support
Avere vFXT instances can now use the primary and secondary IP ranges configured with the subnet for the cluster addresses.

NOTE that Avere vFXT deployment tools try to validate that any addresses are not used elsewhere.  This is not possible for Shared VPC configurations that export a subnetwork into a different project and share the same address space.  The vFXT deployment could choose an in use address if no addresses are provided.  If this happens the following error will be emitted:

    Error: IP '<>/32' is already being used by another resource.