msazure: nic ip configuration add/remove reworking
Retry add/remove of an ip configuration 60 times, but refresh the nic data on
error in case we get out of date.  Since there is no etag visible from
the callers perspective this might prevent spurious stale configurations from
being put until we bubble back up in the error handling.

For operations that are accepted, return errors from the poll (for ex if the
operation was cancelled by the backend) and continues the retry.

If we timeout, then fail the add/remove altogether since that will end up
eating into wall clock pseudo deadline we try and honor.  Increase the poll
time from 180s to 300s.

Also call virtual_networks.check_ip_address_availability as a sanity check
and poll on that for 60s in hopes that any backend book keeping is completed
before we attempt the add of an ip configuration for that address.